requirements.txt
# FastAPI Backend Dependencies (Python 3.14 compatible)
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
gunicorn>=22.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
pydantic>=2.9.0
pydantic-settings>=2.5.0
sqlalchemy>=2.0.30
asyncpg>=0.30.0
psycopg2-binary>=2.9.9
alembic>=1.13.0
redis>=5.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
bcrypt>=4.2.0
PyJWT>=2.9.0
